refactor(posts): extract playlist operations to posts/playlists.ts
Moved all playlist-related operations to a dedicated module with proper JSDoc comments. Fixed type issues and ensured backward compatibility by re-exporting all functions. Added CourseBuilderError class and updated imports in sanity-content-query.ts.
